Refactor JIRA-related modules for better structure.
Moved JIRA value objects to a dedicated module and adjusted imports accordingly for improved organization and separation of concerns. Removed redundant code in previous locations to ensure consistency and maintainability.
